requirements.txt
tensorflow==2.18.0
keras==3.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
